requirements.txt
pandas==2.2.3
gnews==0.4.0
yake==0.4.8
clean-text==0.6.0
newsapi-python==0.2.7
beautifulsoup4==4.12.3
python-dotenv==1.0.1
SQLAlchemy==2.0.38
Unidecode==1.3.8
pymysql==1.1.1